emacs 24 compatibility
`transmission-add' can be troublesome on 24 because it lacks bracketed
paste, so pasting a link with characters like, for instance, a question
mark, is treated as the user hitting a ? in the prompt, which brings up
completions instead of inserting a ?.

The only other compatibility issue is the use of `sort' to sort a
vector. The seq library handles this.

Add post-emacs24 libraries let-alist and seq to Package-Requires.