new commands for adding and removing trackers
`transmission-trackers-add' (bound to `t') used to add new announce
URLs, and `transmission-trackers-remove' (bound to `T') used to remove
trackers by their ID.  These are bound in *transmission*
and *transmission-info* except for `T' in *transmission* since you
cannot see which ID corresponds to which tracker.  This is something
that ought to be improved.

It may be a good idea to instead use `completion-read-multiple' instead
of the current implementation, which is based on what's in the debbugs
package.  However, crm.el is based on completing-read and something more
general will be needed to show more than just a tracker ID in the prompt
when removing a tracker.