give transmission-add some default arguments
transmission-add can now attempt to guess a torrent at point either
by url link text properties, using ffap, dired file names, or the word
at point if it happens to be a BitTorrent info hash. This guess is shown
in the prompt and will be used if an empty input is submitted.