do not errantly send old requests when the session id changes
this would result in, for example, refreshing the torrent list at some
point and the list going blank, then the next refresh brings things
back to normal

this is because the process filter and process plist would only get
cleared from a process used for an asynchronous request _if_ it had a
callback.  so, what would happen is this, as an example:

* delete a torrent, which happens asynchronously and puts the request
  into a process's plist
* the transmission daemon session ID changes.  according to
  libtransmission/session-id.c, this happens every hour.
* M-x transmission or similar:  the first synchronous request gets
  back a 409 and the new session header.  The process filter meant for
  asynchronous requests processes the response and sends a new
  request with content (process-get process :request), which is the
  old delete torrent request.
* transmission responds success, which is basically an empty response
* an empty torrent list is drawn

scary!

so now, transmission-request does like transmission-request-async and
sets the process filter and process plist before doing anything.